working obj reader, mesh class

+++ src/mat4.h
@@ -14,24 +14,24 @@ class mat4 {
 
     public:
 
-        double vals[4][4];
+        float vals[4][4];
 
         mat4();
 
-        void translate(double transX, double transY, double transZ);
-        void scale(double scaleX, double scaleY, double scaleZ);
-        void rotX (double t);
-        void rotY (double t);
-        void rotZ (double t);
+        void translate(float transX, float transY, float transZ);
+        void scale(float scaleX, float scaleY, float scaleZ);
+        void rotX (float t);
+        void rotY (float t);
+        void rotZ (float t);
 
-        double & operator()(const int i, const int j);
-        double operator()(const int i, const int j) const;
+        float & operator()(const int i, const int j);
+        float operator()(const int i, const int j) const;
 
         mat4 & operator*=(const mat4& m);
-        mat4 & operator*=(const double c);
+        mat4 & operator*=(const float c);
 
         friend mat4 operator*(const mat4& m1, mat4& m2);
-        friend mat4 operator*(const mat4&m, const double c);
+        friend mat4 operator*(const mat4&m, const float c);
 
 };
