+++ src/vec4.cpp
@@ -2,14 +2,14 @@
 
 // empty constructor
 vec4::vec4() {
-	x = 0.0;
-	y = 0.0;
-	z = 0.0;
-	w = 1.0;
+	x = 0.0f;
+	y = 0.0f;
+	z = 0.0f;
+	w = 1.0f;
 }
 
 // initialize entire vector with one value
-vec4::vec4(double a) {
+vec4::vec4(float a) {
 	x = a;
 	y = a;
 	z = a;
@@ -17,21 +17,21 @@ vec4::vec4(double a) {
 }
 
 // basic constructor
-vec4::vec4(double xx, double yy, double zz) {
+vec4::vec4(float xx, float yy, float zz) {
 	x = xx;
 	y = yy;
 	z = zz;
-	w = 1.0;
+	w = 1.0f;
 }
 
 // includes w value
-vec4::vec4(double xx, double yy, double zz, double ww) {
+vec4::vec4(float xx, float yy, float zz, float ww) {
 	// check if w can divide the other values
-	if (ww != 0) {
+	if (ww != 0.0f) {
 		x = xx / ww;
 		y = yy / ww;
 		z = zz / ww;
-		w = 1.0;
+		w = 1.0f;
 	}
 	else {
 		x = xx;
@@ -42,33 +42,33 @@ vec4::vec4(double xx, double yy, double zz, double ww) {
 }
 
 // magnitude of vector
-double vec4::length() const {
+float vec4::length() const {
 	return sqrt((x*x) + (y*y) + (z*z));
 }
 
 // dot product of vector
-double vec4::dot(vec4& vec) {
+float vec4::dot(vec4& vec) {
 	return (x * vec.x) + (y * vec.y) + (z * vec.z);
 }
 
 // cross product of vector
 vec4 vec4::cross(vec4& vec) {
-	double newx = (y * vec.z) - (z * vec.y);
-	double newy = (z * vec.x) - (x * vec.z);
-	double newz = (x * vec.y) - (y * vec.x);
+	float newx = (y * vec.z) - (z * vec.y);
+	float newy = (z * vec.x) - (x * vec.z);
+	float newz = (x * vec.y) - (y * vec.x);
 	return vec4(newx, newy, newz);
 }
 
 // normalize a vector
 void vec4::normalize() {
-	double len = length();
+	float len = length();
 	x /= len;
 	y /= len;
 	z /= len;
 }
 
 // clamps the values of the vector
-void vec4::clamp(double low, double high) {
+void vec4::clamp(float low, float high) {
 	if (x < low) {
 		x = low;
 	}
@@ -135,29 +135,29 @@ vec4& vec4::operator-=(const vec4& v) {
 	return *this;
 }
 
-vec4& vec4::operator=(const double c) {
+vec4& vec4::operator=(const float c) {
 	x = c;
 	y = c;
 	z = c;
-	w = 1.0;
+	w = 1.0f;
 	return *this;
 }
 
-vec4& vec4::operator+=(const double c) {
+vec4& vec4::operator+=(const float c) {
 	x += c;
 	y += c;
 	z += c;
 	return *this;
 }
 
-vec4& vec4::operator-=(const double c) {
+vec4& vec4::operator-=(const float c) {
 	x -= c;
 	y -= c;
 	z -= c;
 	return *this;
 }
 
-vec4& vec4::operator*=(const double c) {
+vec4& vec4::operator*=(const float c) {
 	x *= c;
 	y *= c;
 	z *= c;
@@ -172,19 +172,19 @@ vec4 operator+(const vec4& a, const vec4& b) {
 	return vec4(a.x + b.x, a.y + b.y, a.z + b.z);
 }
 
-vec4 operator-(const vec4& a, const double c) {
+vec4 operator-(const vec4& a, const float c) {
 	return vec4(a.x - c, a.y - c, a.z - c);
 }
 
-vec4 operator-(const double c, const vec4& a) {
+vec4 operator-(const float c, const vec4& a) {
 	return vec4(a.x - c, a.y - c, a.z - c);
 }
 
-vec4 operator+(const vec4& a, const double c) {
+vec4 operator+(const vec4& a, const float c) {
 	return vec4(a.x + c, a.y + c, a.z + c);
 }
 
-vec4 operator+(const double c, const vec4& a) {
+vec4 operator+(const float c, const vec4& a) {
 	return vec4(a.x + c, a.y + c, a.z + c);
 }
 
@@ -192,11 +192,11 @@ vec4 operator*(const vec4& a, const vec4& b) {
 	return vec4(a.x * b.x, a.y * b.y, a.z * b.z);
 }
 
-vec4 operator*(const double c, const vec4& a) {
+vec4 operator*(const float c, const vec4& a) {
 	return vec4(a.x * c, a.y * c, a.z * c);
 }
 
-vec4 operator*(const vec4& a, const double c) {
+vec4 operator*(const vec4& a, const float c) {
 	return vec4(a.x * c, a.y * c, a.z * c);
 }
